+++ src/helm/benchmark/metrics/codeinsights_metric_specs.py
@@ -5,7 +5,7 @@ from helm.benchmark.metrics.metric import MetricSpec
 def get_functional_correctness_metric_specs(compile_code: bool = True) -> List[MetricSpec]:
     return [
         MetricSpec(
-            class_name="helm.benchmark.metrics.correct_code_metrics.FunctionalCorrectnessMetric",
+            class_name="helm.benchmark.metrics.codeinsights_correct_code_metrics.FunctionalCorrectnessMetric",
             args={"compile_code": compile_code},
         )
     ]
@@ -16,31 +16,33 @@ def get_comprehensive_code_evaluation_metric_specs(
 ) -> List[MetricSpec]:
     return [
         MetricSpec(
-            class_name="helm.benchmark.metrics.code_evaluation_metrics.ComprehensiveCodeEvaluationMetric",
+            class_name="helm.benchmark.metrics.codeinsights_code_evaluation_metrics.ComprehensiveCodeEvaluationMetric",
             args={"use_codebert": use_codebert, "compile_code": compile_code},
         )
     ]
+
 
 def get_code_efficiency_metric_specs(
     compile_code: bool = True,
     num_runtime_runs: int = 5,
     timeout_seconds: int = 10,
-    use_codebert: bool = True,               # ➊ add arg if you wish
+    use_codebert: bool = True,  # ➊ add arg if you wish
 ):
     return [
-        MetricSpec(                             # existing metric → runtime & correctness
-            class_name="helm.benchmark.metrics.code_efficiency_metrics.CodeEfficiencyMetric",
+        MetricSpec(  # existing metric → runtime & correctness
+            class_name="helm.benchmark.metrics.codeinsights_code_efficiency_metrics.CodeEfficiencyMetric",
             args={
                 "compile_code": compile_code,
                 "num_runtime_runs": num_runtime_runs,
                 "timeout_seconds": timeout_seconds,
             },
         ),
-        MetricSpec(                             # ➋ NEW metric → AST + CodeBERT
-            class_name="helm.benchmark.metrics.code_evaluation_metrics.CodeEvaluationMetric",
+        MetricSpec(  # ➋ NEW metric → AST + CodeBERT
+            class_name="helm.benchmark.metrics.codeinsights_code_evaluation_metrics.CodeEvaluationMetric",
             args={"use_codebert": use_codebert},
         ),
     ]
+
 
 def get_edge_case_metric_specs(
     use_codebert: bool = True,
@@ -49,7 +51,7 @@ def get_edge_case_metric_specs(
 ) -> List[MetricSpec]:
     return [
         MetricSpec(
-            class_name="helm.benchmark.metrics.edge_case_metrics.ComprehensiveCodeEvaluationMetric",
+            class_name="helm.benchmark.metrics.codeinsights_edge_case_metrics.ComprehensiveCodeEvaluationMetric",
             args={
                 "use_codebert": use_codebert,
                 "compile_code": compile_code,
