+++ src/helm/benchmark/presentation/run_entries_codeinsights.conf
@@ -3,5 +3,7 @@
 entries: [
   {description: "correct_code:model=text", priority: 1}
   {description: "student_coding:model=text", priority: 1}
-  {description: "student_mistake_coding:model=text", priority: 1}
+  {description: "student_mistake:model=text", priority: 1}
+  {description: "edge_case:model=text", priority: 1}
+  {description: "code_efficiency:model=text", priority: 1}
 ]
